+++ src/app.rs
@@ -1,13 +1,9 @@
 use std::fs;
 use std::path::{Path, PathBuf};
 
-use log::info;
 use rand::rngs::OsRng;
 use rand::RngCore;
-use tokio::fs::try_exists;
-use tokio::task::spawn_blocking;
 
-use crate::app::legacy::*;
 use crate::paths::Paths;
 
 pub mod auth;
@@ -15,7 +11,6 @@ pub mod config;
 pub mod ddns;
 pub mod formats;
 pub mod index;
-pub mod legacy;
 pub mod ndb;
 pub mod peaks;
 pub mod playlist;
@@ -36,8 +31,6 @@ pub enum Error {
 	Io(PathBuf, std::io::Error),
 	#[error(transparent)]
 	FileWatch(#[from] notify::Error),
-	#[error(transparent)]
-	SQL(#[from] rusqlite::Error),
 	#[error(transparent)]
 	Ape(#[from] ape::Error),
 	#[error("ID3 error in `{0}`: `{1}`")]
@@ -179,7 +172,6 @@ impl App {
 			.map_err(|e| Error::Io(thumbnails_dir_path.clone(), e))?;
 
 		let auth_secret_file_path = paths.data_dir_path.join("auth.secret");
-		Self::migrate_legacy_auth_secret(&paths.db_file_path, &auth_secret_file_path).await?;
 		let auth_secret = Self::get_or_create_auth_secret(&auth_secret_file_path).await?;
 
 		let config_manager = config::Manager::new(&paths.config_file_path, auth_secret).await?;
@@ -203,98 +195,7 @@ impl App {
 			thumbnail_manager,
 		};
 
-		app.migrate_legacy_db(&paths.db_file_path).await?;
-
 		Ok(app)
-	}
-
-	async fn migrate_legacy_auth_secret(
-		db_file_path: &PathBuf,
-		secret_file_path: &PathBuf,
-	) -> Result<(), Error> {
-		if !try_exists(db_file_path)
-			.await
-			.map_err(|e| Error::Io(db_file_path.clone(), e))?
-		{
-			return Ok(());
-		}
-
-		if try_exists(secret_file_path)
-			.await
-			.map_err(|e| Error::Io(secret_file_path.clone(), e))?
-		{
-			return Ok(());
-		}
-
-		info!(
-			"Migrating auth secret from database at `{}`",
-			db_file_path.to_string_lossy()
-		);
-
-		let secret = spawn_blocking({
-			let db_file_path = db_file_path.clone();
-			move || read_legacy_auth_secret(&db_file_path)
-		})
-		.await??;
-
-		tokio::fs::write(secret_file_path, &secret)
-			.await
-			.map_err(|e| Error::Io(secret_file_path.clone(), e))?;
-
-		Ok(())
-	}
-
-	async fn migrate_legacy_db(&self, db_file_path: &PathBuf) -> Result<(), Error> {
-		if !try_exists(db_file_path)
-			.await
-			.map_err(|e| Error::Io(db_file_path.clone(), e))?
-		{
-			return Ok(());
-		}
-
-		let Some(config) = tokio::task::spawn_blocking({
-			let db_file_path = db_file_path.clone();
-			move || read_legacy_config(&db_file_path)
-		})
-		.await??
-		else {
-			return Ok(());
-		};
-
-		info!(
-			"Found usable config in legacy database at `{}`, beginning migration process",
-			db_file_path.to_string_lossy()
-		);
-
-		info!("Migrating configuration");
-		self.config_manager.apply_config(config).await?;
-		self.config_manager.save_config().await?;
-
-		info!("Migrating playlists");
-		for (name, owner, songs) in read_legacy_playlists(
-			db_file_path,
-			self.index_manager.clone(),
-			self.scanner.clone(),
-		)
-		.await?
-		{
-			self.playlist_manager
-				.save_playlist(&name, &owner, songs)
-				.await?;
-		}
-
-		info!(
-			"Deleting legacy database at `{}`",
-			db_file_path.to_string_lossy()
-		);
-		delete_legacy_db(db_file_path).await?;
-
-		info!(
-			"Completed migration from `{}`",
-			db_file_path.to_string_lossy()
-		);
-
-		Ok(())
 	}
 
 	async fn get_or_create_auth_secret(path: &Path) -> Result<auth::Secret, Error> {