+++ src/main.rs
@@ -129,7 +129,6 @@ fn main() -> Result<(), Error> {
 	info!("Cache files location is {:#?}", paths.cache_dir_path);
 	info!("Data files location is {:#?}", paths.data_dir_path);
 	info!("Config file location is {:#?}", paths.config_file_path);
-	info!("Legacy database file location is {:#?}", paths.db_file_path);
 	info!("Log file location is {:#?}", paths.log_file_path);
 	#[cfg(unix)]
 	if !cli_options.foreground {