+++ src/options.rs
@@ -8,7 +8,6 @@ pub struct CLIOptions {
 	#[cfg(unix)]
 	pub pid_file_path: Option<PathBuf>,
 	pub config_file_path: Option<PathBuf>,
-	pub database_file_path: Option<PathBuf>,
 	pub cache_dir_path: Option<PathBuf>,
 	pub data_dir_path: Option<PathBuf>,
 	pub web_dir_path: Option<PathBuf>,
@@ -40,7 +39,6 @@ impl Manager {
 			#[cfg(unix)]
 			pid_file_path: matches.opt_str("pid").map(PathBuf::from),
 			config_file_path: matches.opt_str("c").map(PathBuf::from),
-			database_file_path: matches.opt_str("d").map(PathBuf::from),
 			cache_dir_path: matches.opt_str("cache").map(PathBuf::from),
 			data_dir_path: matches.opt_str("data").map(PathBuf::from),
 			web_dir_path: matches.opt_str("w").map(PathBuf::from),