+++ src/paths.rs
@@ -6,7 +6,6 @@ pub struct Paths {
 	pub cache_dir_path: PathBuf,
 	pub config_file_path: PathBuf,
 	pub data_dir_path: PathBuf,
-	pub db_file_path: PathBuf,
 	pub log_file_path: Option<PathBuf>,
 	#[cfg(unix)]
 	pub pid_file_path: PathBuf,
@@ -22,7 +21,6 @@ impl Default for Paths {
 			cache_dir_path: ["."].iter().collect(),
 			config_file_path: [".", "polaris.toml"].iter().collect(),
 			data_dir_path: ["."].iter().collect(),
-			db_file_path: [".", "db.sqlite"].iter().collect(),
 			log_file_path: Some([".", "polaris.log"].iter().collect()),
 			pid_file_path: [".", "polaris.pid"].iter().collect(),
 			web_dir_path: [".", "web"].iter().collect(),
@@ -40,7 +38,6 @@ impl Default for Paths {
 			cache_dir_path: install_directory.clone(),
 			config_file_path: install_directory.join("polaris.toml"),
 			data_dir_path: install_directory.clone(),
-			db_file_path: install_directory.join("db.sqlite"),
 			log_file_path: Some(install_directory.join("polaris.log")),
 			web_dir_path: install_directory.join("web"),
 		}
@@ -51,10 +48,6 @@ impl Paths {
 	fn from_build() -> Self {
 		let defaults = Self::default();
 		Self {
-			db_file_path: option_env!("POLARIS_DB_DIR")
-				.map(PathBuf::from)
-				.map(|p| p.join("db.sqlite"))
-				.unwrap_or(defaults.db_file_path),
 			cache_dir_path: option_env!("POLARIS_CACHE_DIR")
 				.map(PathBuf::from)
 				.unwrap_or(defaults.cache_dir_path),
@@ -89,9 +82,6 @@ impl Paths {
 		}
 		if let Some(path) = &cli_options.data_dir_path {
 			path.clone_into(&mut paths.data_dir_path);
-		}
-		if let Some(path) = &cli_options.database_file_path {
-			path.clone_into(&mut paths.db_file_path);
 		}
 		#[cfg(unix)]
 		if let Some(path) = &cli_options.pid_file_path {