+++ src/server/axum/test.rs
@@ -25,7 +25,6 @@ impl TestService for AxumTestService {
 			cache_dir_path: ["test-output", test_name].iter().collect(),
 			config_file_path: output_dir.join("polaris.toml"),
 			data_dir_path: ["test-output", test_name].iter().collect(),
-			db_file_path: output_dir.join("db.sqlite"),
 			#[cfg(unix)]
 			pid_file_path: output_dir.join("polaris.pid"),
 			log_file_path: None,